// src/app/modules/customers/customer.type.ts
export type ICustomerFilterRequest = {
  searchTerm?: string | undefined
  firstName?: string | undefined
  middleName?: string | undefined
  lastName?: string | undefined
  uniqueId?: string | undefined
  role?: string | undefined
  nid?: string | undefined
  email?: string | undefined
  phoneNo?: string | undefined
  gender?: string | undefined
  presentAddress?: string | undefined
  permanentAddress?: string | undefined
}